+++ src/lib/assertEnsures.ts
@@ -12,19 +12,19 @@ import unChecked from './unChecked';
 /**
  * Applies invariant assertions to the provided class instance
  *
- * @param {Record<PropertyKey, unknown>} ctx - The context class
+ * @param {U} ctx - The context class
  * @param {Contract<any>} contract - The contract
  * @param {string} className - The name of the class
  * @param {PropertyKey} featureName - The name of the feature
- * @param {Record<PropertyKey, unknown>} old - The original properties before execution
+ * @param {U} old - The original properties before execution
  * @param {any[]} args - The arguments of the feature to apply to the assertion
  */
-function assertEnsures(
-    ctx: Record<PropertyKey, unknown>,
+function assertEnsures<U>(
+    ctx: U,
     contract: Contract<any>,
     className: string,
     featureName: PropertyKey,
-    old: Record<PropertyKey, unknown>,
+    old: U,
     args: any[]
 ){
     const e: Ensures<any, any> | undefined = Reflect.get(contract.assertions, featureName)?.ensures,