+++ src/lib/assertInvariants.ts
@@ -11,10 +11,10 @@ import { assert, checkedMode, Contract, extend, invariant } from '../';
 /**
  * Applies invariant assertions to the provided class instance
  *
- * @param {Record<PropertyKey, unknown>} ctx - The context class
+ * @param {U} ctx - The context class
  * @param {Contract<any>} contract - The contract
  */
-function assertInvariants(ctx: Record<PropertyKey, unknown>, contract: Contract<any>) {
+function assertInvariants<U>(ctx: U, contract: Contract<any>) {
     if(contract[checkedMode]) {
         unChecked(contract,() => {
             const iv = contract[invariant];